Support versions older than 1.19
The launcher can now download and run versions older than 1.19 but
newer than 1.12.2.
Versions older than 1.19 have a 'classifiers' entry in their library
json. This file also needs to be downloaded along with the other
libraries based on some conditions(Operating System). Classifiers are
later extracted to the instances natives directory when the instance is created