Instance manager now spawns a thread for emitting logs
The child process can now emit logs to the frontend using the
app_handle. The instance manager spawns a new thread for the running
instance that continually reads stdout.

The frontend can now display the logs emitted from an instance and
display them in a list on the "logs" tab. Logs are stored on the main
'Home' page and passed to the Logs component. They will need to be
cleared when an instance is closed.